obtain plots from spice and modify spice netlists, AT4 improved
So much stuff...
Also, don't trust function LTspice2Matlab. I got from the internet and it's not the most solid function in the world. I had to modify it a bit, and it seems to work only on ascii-formatted .raw files. I'll try to fix a bit more next time